requirements.txt
kopf
terminal-shop
python-dotenv
kubernetes